@@ -3,10 +3,11 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "react": "^15.4.2",
+    "prop-types": "^15.5.8",
+    "react": "^15.5.4",
     "react-addons-create-fragment": "^15.5.3",
     "react-apollo": "^1.0.1",
-    "react-dom": "^15.4.2"
+    "react-dom": "^15.5.4"
   },
   "devDependencies": {
     "react-scripts": "0.9.5"
